requirements.txt
# Core dependencies
selenium==4.10.0
python-dotenv==1.0.0
webdriver-manager==3.8.6
requests==2.31.0
pillow==10.0.0
loguru==0.7.0

# Email handling
imap-tools==1.0.0
email-validator==2.0.0

# Captcha solving
2captcha-python==1.1.3
pytesseract==0.3.10

# Utilities
opencv-python==4.8.0.74
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2

# Testing
pytest==7.4.0